refactor(DB): Moathon 모델을 accounts에서 challenges 앱으로 분리 및 이관
- 서비스 핵심 도메인인 Moathon을 독립적인 challenges 앱으로 이동하여 아키텍처 개선

- accounts.User 모델의 products 필드 through 경로 수정 ('challenges.Moathon')

- 데이터 시딩 스크립트(seed_fake_users_moathon.py) 및 View/Serializer의 Import 경로 수정

- 모델 이동에 따른 기존 마이그레이션 파일 정리 및 재생성
